Animated tiles, cutscene controller
Added and placed new room lava tiles tiles. Removed the chain that was placed in the room next to the cage and now have the cage spawning it when created. Fixed a bug with the rock worms where depending on the rotation of the object the projectile would spawn in the wall. Updated the projectile to add force depending on the side of the enemy that it hits. Created a few more objects to use in the cutscene room and updated the level room to Room_LevelOne