More cutscene changes, endZone object, floor
Creation a floor object and modifications to enemy object so it can be used without both scales in the room. Changes to the target room variable for the transition parent. Added walls into cutscene room and modified a few background tiles. Changes to the player pawn's x location when in the cage and added x scale rotation to match the players cursor.